overwrite original file
